Removed Ganon BK on LACS but add Dungeon and Keysanity options
Ganon BK balancing

BK on LACS is hard to balance without a ton of conditions since it is useless in 6 medaillons bridge or all dungeons.
Dungeon asks to search trials for checks so it may lead to fun endings !
Keysanity just adds another item to find anywhere in the world.